requirements.txt
openai==0.28.0
psycopg2
sentence-transformers
openai
python-dotenv

